A Sound Decision
An experimental commit playing with sounds and adding songs to both the jukebox to all maps and to the lobby music of Mollusc ONLY.

The update also adds fire extinguishers to the sparse selection in the station's stretched out hallways with markings to help them stand out.

Sounds of select types of weapons have been replaced for punchier sounds, this WILL have mixed reactions and may be undone. Time will tell.

Functionality has been added to the grenade launcher of the Yinglet Adapted Rifle, no longer will it eat a grenade and refuse to spit it back out.