Remove the ability for baxxid to use guns/firearms.
An oversight when altering the baxxid dexterity is that they could now use guns and other firearms, which unbalances them compared to other ingame species. AND-NOT-ing DEXTERITY_FULL with DEXTERITY_WEAPONS appears to fix this issue. Melee tools such as the crowbar or katana seem to still function from testing it on a private server.