Fix surgery checking surgeon's clothing for blockage instead of patient's
Fixes an issue where if the surgeon was wearing a thick material on a body part, they would be unable to perform surgery on a patient even if the patient wasn't wearing anything.

Also added to the failure message a section that explains what item is blocking the surgery, removing guesswork.