Last try for a few hours
final go before I go insane
if this doesn't work, I'm manually adding every item's scripting into summit.dm and that's going to be it
#include causes a runtime
not #include-ing causes atoms to be missing
and zat_zing basically means I'm adding the entire Yinglet species to summit.dm JUST to have it on the map as an easter egg. I truly am living the life here. lol